Change `Rng_Rand16` to return `s32`

+++ include/maps/shared/Ai_HangedScratcher_Init.h
@@ -118,8 +118,8 @@ void Ai_HangedScratcher_Init(s_SubCharacter* chara)
     }
 
     // TODO: Use `Rng_GenerateInt`.
-    chara->properties_E4.larvalStalker.properties_E8[9].val16[0] = ((s32)Rng_Rand16() % (var_s0 >> 2)) + ((var_s0 * 3) >> 2);
-    chara->properties_E4.larvalStalker.properties_E8[9].val16[1] = ((s32)Rng_Rand16() % (var_s2 >> 2)) + ((var_s2 * 3) >> 2);
+    chara->properties_E4.larvalStalker.properties_E8[9].val16[0] = (Rng_Rand16() % (var_s0 >> 2)) + ((var_s0 * 3) >> 2);
+    chara->properties_E4.larvalStalker.properties_E8[9].val16[1] = (Rng_Rand16() % (var_s2 >> 2)) + ((var_s2 * 3) >> 2);
 
     chara->flags_3E |= CharaFlag_Unk9 | CharaFlag_Unk3;
 }
